server/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.11.0
django-redis==5.2.0
djangorestframework==3.13.1
engineering-notation==0.6.0
httpie==3.1.0
idna==3.3
multidict==6.0.2
packaging==21.3
Pygments==2.11.2
pyparsing==3.0.7
PySocks==1.7.1
pytz==2022.1
PyYAML==6.0
redis==4.2.0
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
tk==0.1.0
tk-tools==0.14.0
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0